refactor to be consistent with other transient saves. add mutex
